Convert test-manager config into a module
