+++ test/test-manager/src/config/vm.rs
@@ -1,141 +1,9 @@
-//! Test manager configuration.
+//! Virtual machine configuration.
+
+use std::env;
+use std::path::{Path, PathBuf};
 
 use serde::{Deserialize, Serialize};
-use std::{
-    collections::BTreeMap,
-    env, io,
-    ops::Deref,
-    path::{Path, PathBuf},
-};
-
-use crate::tests::config::DEFAULT_MULLVAD_HOST;
-
-#[derive(thiserror::Error, Debug)]
-pub enum Error {
-    #[error("Could not find config dir")]
-    FindConfigDir,
-    #[error("Could not create config dir")]
-    CreateConfigDir(#[source] io::Error),
-    #[error("Failed to read config")]
-    Read(#[source] io::Error),
-    #[error("Failed to parse config")]
-    InvalidConfig(#[from] serde_json::Error),
-    #[error("Failed to write config")]
-    Write(#[source] io::Error),
-}
-
-#[derive(Default, Serialize, Deserialize, Clone)]
-pub struct Config {
-    #[serde(skip)]
-    pub runtime_opts: RuntimeOptions,
-    pub vms: BTreeMap<String, VmConfig>,
-    pub mullvad_host: Option<String>,
-}
-
-#[derive(Default, Serialize, Deserialize, Clone)]
-pub struct RuntimeOptions {
-    pub display: Display,
-    pub keep_changes: bool,
-}
-
-#[derive(Default, Serialize, Deserialize, Clone)]
-pub enum Display {
-    #[default]
-    None,
-    Local,
-    Vnc,
-}
-
-impl Config {
-    async fn load_or_default<P: AsRef<Path>>(path: P) -> Result<Self, Error> {
-        Self::load(path).await.or_else(|error| match error {
-            Error::Read(ref io_err) if io_err.kind() == io::ErrorKind::NotFound => {
-                Ok(Self::default())
-            }
-            error => Err(error),
-        })
-    }
-
-    async fn load<P: AsRef<Path>>(path: P) -> Result<Self, Error> {
-        let data = tokio::fs::read(path).await.map_err(Error::Read)?;
-        serde_json::from_slice(&data).map_err(Error::InvalidConfig)
-    }
-
-    async fn save<P: AsRef<Path>>(&self, path: P) -> Result<(), Error> {
-        let data = serde_json::to_vec_pretty(self).unwrap();
-        tokio::fs::write(path, &data).await.map_err(Error::Write)
-    }
-
-    pub fn get_vm(&self, name: &str) -> Option<&VmConfig> {
-        self.vms.get(name)
-    }
-
-    /// Get the Mullvad host to use.
-    ///
-    /// Defaults to [`DEFAULT_MULLVAD_HOST`] if the host was not provided in the [`ConfigFile`].
-    pub fn get_host(&self) -> String {
-        self.mullvad_host.clone().unwrap_or_else(|| {
-            log::debug!("No Mullvad host has been set explicitly. Falling back to default host");
-            DEFAULT_MULLVAD_HOST.to_owned()
-        })
-    }
-}
-
-pub struct ConfigFile {
-    path: PathBuf,
-    config: Config,
-}
-
-impl ConfigFile {
-    /// Make config changes and save them to disk
-    pub async fn load_or_default() -> Result<Self, Error> {
-        Self::load_or_default_inner(Self::get_config_path()?).await
-    }
-
-    /// Get configuration file path
-    fn get_config_path() -> Result<PathBuf, Error> {
-        Ok(Self::get_config_dir()?.join("config.json"))
-    }
-
-    /// Get configuration file directory
-    fn get_config_dir() -> Result<PathBuf, Error> {
-        let dir = dirs::config_dir()
-            .ok_or(Error::FindConfigDir)?
-            .join("mullvad-test");
-        Ok(dir)
-    }
-
-    /// Create configuration file directory if it does not exist
-    async fn ensure_config_dir() -> Result<(), Error> {
-        tokio::fs::create_dir_all(Self::get_config_dir()?)
-            .await
-            .map_err(Error::CreateConfigDir)
-    }
-
-    /// Make config changes and save them to disk
-    async fn load_or_default_inner<P: AsRef<Path>>(path: P) -> Result<Self, Error> {
-        Ok(Self {
-            path: path.as_ref().to_path_buf(),
-            config: Config::load_or_default(path).await?,
-        })
-    }
-
-    /// Make config changes and save them to disk
-    pub async fn edit(&mut self, edit: impl FnOnce(&mut Config)) -> Result<(), Error> {
-        Self::ensure_config_dir().await?;
-
-        edit(&mut self.config);
-        self.config.save(&self.path).await
-    }
-}
-
-impl Deref for ConfigFile {
-    type Target = Config;
-
-    fn deref(&self) -> &Self::Target {
-        &self.config
-    }
-}
 
 #[derive(clap::Args, Debug, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "snake_case")]
